+++ scratch.js
@@ -7,7 +7,7 @@ function from_file(page) {
   let str = require('fs').readFileSync('./tests/cache/' + page.toLowerCase() + '.txt', 'utf-8');
   let options = {};
   let r = wtf(str, options);
-  console.log(r.toPlaintext());
+  console.log(r.infoboxes());
 }
 
 // wtf.from_api('Aldous Huxley', 'en', function(markup) {
@@ -19,8 +19,8 @@ function from_file(page) {
 // from_file('earthquakes');
 // from_file('al_Haytham');
 // from_file('redirect');
-// from_file('raith_rovers');
-from_file('royal_cinema');
+from_file('raith_rovers');
+// from_file('royal_cinema');
 // from_file('Toronto_Star');
 // from_file('Toronto');
 // from_file('royal_cinema');
