+++ src/document/parse/section/image/index.js
@@ -1,11 +1,8 @@
 const i18n = require('../../../../data/i18n');
-const find_recursive = require('../../lib/recursive_match');
-const parse_image = require('./image');
+const parse_image = require('./parse-image');
 const fileRegex = new RegExp('(' + i18n.images.concat(i18n.files).join('|') + '):.*?[\\|\\]]', 'i');
 
-const parseImages = function(r, wiki, options) {
-  //second, remove [[file:...[[]] ]] recursions
-  let matches = find_recursive('[', ']', wiki);
+const parseImages = function(matches, r, wiki, options) {
   matches.forEach(function(s) {
     if (s.match(fileRegex)) {
       r.images = r.images || [];
@@ -16,18 +13,7 @@ const parseImages = function(r, wiki, options) {
     }
   });
 
-  //third, wiktionary-style interlanguage links
-  matches.forEach(function(s) {
-    if (s.match(/\[\[([a-z]+):(.*?)\]\]/i) !== null) {
-      let site = (s.match(/\[\[([a-z]+):/i) || [])[1] || '';
-      site = site.toLowerCase();
-      if (site && i18n.dictionary[site] === undefined && !(options.namespace !== undefined && options.namespace === site)) {
-        r.interwiki = r.interwiki || {};
-        r.interwiki[site] = (s.match(/\[\[([a-z]+):(.*?)\]\]/i) || [])[2];
-        wiki = wiki.replace(s, '');
-      }
-    }
-  });
+
   return wiki;
 };
 module.exports = parseImages;
